+++ src/api/api.h
@@ -145,6 +145,7 @@ int luax_readscale(lua_State* L, int index, float* v, int components, const char
 int luax_readquat(lua_State* L, int index, float* q, const char* expected);
 int luax_readmat4(lua_State* L, int index, float* m, int scaleComponents);
 void luax_pushvec3(lua_State* L, float v[3], bool tableArray);
+bool luax_isquat(lua_State* L, int index);
 
 // Module helpers