+++ src/api/l_math_mat4.c
@@ -57,6 +57,15 @@ static int l_lovrMat4GetPosition(lua_State* L) {
   return 3;
 }
 
+static int l_lovrMat4SetPosition(lua_State* L) {
+  float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
+  float position[3];
+  luax_readvec3(L, 2, position, NULL);
+  mat4_setPosition(m, position);
+  lua_settop(L, 1);
+  return 1;
+}
+
 static int l_lovrMat4GetOrientation(lua_State* L) {
   float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
   float angle, ax, ay, az;
@@ -68,6 +77,15 @@ static int l_lovrMat4GetOrientation(lua_State* L) {
   return 4;
 }
 
+static int l_lovrMat4SetOrientation(lua_State* L) {
+  float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
+  float orientation[4];
+  luax_readquat(L, 2, orientation, NULL);
+  mat4_setOrientation(m, orientation);
+  lua_settop(L, 1);
+  return 1;
+}
+
 static int l_lovrMat4GetScale(lua_State* L) {
   float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
   float scale[3];
@@ -76,6 +94,15 @@ static int l_lovrMat4GetScale(lua_State* L) {
   lua_pushnumber(L, scale[1]);
   lua_pushnumber(L, scale[2]);
   return 3;
+}
+
+static int l_lovrMat4SetScale(lua_State* L) {
+  float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
+  float scale[3];
+  luax_readvec3(L, 2, scale, NULL);
+  mat4_setScale(m, scale);
+  lua_settop(L, 1);
+  return 1;
 }
 
 static int l_lovrMat4GetPose(lua_State* L) {
@@ -91,6 +118,18 @@ static int l_lovrMat4GetPose(lua_State* L) {
   lua_pushnumber(L, ay);
   lua_pushnumber(L, az);
   return 7;
+}
+
+static int l_lovrMat4SetPose(lua_State* L) {
+  float* m = lovrMat4GetData(luax_checktype(L, 1, Mat4));
+  float pos[3];
+  float orientation[4];
+  int index = luax_readvec3(L, 2, pos, NULL);
+  luax_readquat(L, index, orientation, NULL);
+  mat4_setPosition(m, pos);
+  mat4_setOrientation(m, orientation);
+  lua_settop(L, 1);
+  return 1;
 }
 
 int l_lovrMat4Set(lua_State* L) {
@@ -114,17 +153,17 @@ int l_lovrMat4Set(lua_State* L) {
     mat4_identity(m);
 
     float position[3], orientation[4], scale[3];
-    index = luax_readvec3(L, index, position, "nil, number, vec3, or mat4");
+    index = luax_readvec3(L, index, position, "nil, number, table, vector, or Mat4");
     m[12] = position[0];
     m[13] = position[1];
     m[14] = position[2];
 
     // 1 more arg or 4 numbers: rotation, otherwise scale + rotation
-    if (top == index || ((top - index) == 3 && lua_type(L, top) == LUA_TNUMBER)) {
+    if (luax_isquat(L, index) || ((top - index) == 3 && lua_type(L, top) == LUA_TNUMBER)) {
       luax_readquat(L, index, orientation, NULL);
       mat4_rotateQuat(m, orientation);
     } else {
-      index = luax_readscale(L, index, scale, 3, NULL);
+      index = luax_readscale(L, index, scale, 3, "nil, number, table, vector, or quaternion");
       index = luax_readquat(L, index, orientation, NULL);
       mat4_rotateQuat(m, orientation);
       mat4_scale(m, scale[0], scale[1], scale[2]);
@@ -160,11 +199,12 @@ static int l_lovrMat4Mul(lua_State* L) {
     mat4_mulVec4(lovrMat4GetData(matrix), v);
     if (lua_istable(L, 2)) {
       luax_pushvec3(L, v, luax_len(L, 2) > 0);
+      if (lua_getmetatable(L, 2)) {
+        lua_setmetatable(L, -2);
+      }
     } else {
 #ifdef LOVR_USE_LUAU
       lua_pushvector(L, v[0], v[1], v[2]);
-#else
-      luax_pushvec3(L, v, false);
 #endif
     }
     return 1;
@@ -314,7 +354,7 @@ static int l_lovrMat4__mul(lua_State* L) {
   if (other) {
     Mat4* result = lovrMat4Create();
     mat4_init(lovrMat4GetData(result), lovrMat4GetData(self));
-    mat4_mul(lovrMat4GetData(self), lovrMat4GetData(other));
+    mat4_mul(lovrMat4GetData(result), lovrMat4GetData(other));
     luax_pushtype(L, Mat4, result);
     lovrRelease(result, lovrMat4Destroy);
     return 1;
@@ -325,6 +365,9 @@ static int l_lovrMat4__mul(lua_State* L) {
     luax_readvec3(L, 2, v, NULL);
     mat4_mulPoint(lovrMat4GetData(self), v);
     luax_pushvec3(L, v, luax_len(L, 2) > 0);
+    if (lua_getmetatable(L, 2)) {
+      lua_setmetatable(L, -2);
+    }
     return 1;
   }
 
@@ -339,7 +382,7 @@ static int l_lovrMat4__mul(lua_State* L) {
   }
 #endif
 
-  return luaL_error(L, "Bad right hand side for Mat4 * operator: expected a Mat4 or a vector");
+  return luaL_error(L, "Bad right hand side for Mat4 * operator: expected a Mat4, table, or vector");
 }
 
 static int l_lovrMat4__tostring(lua_State* L) {
@@ -365,7 +408,7 @@ static int l_lovrMat4__newindex(lua_State* L) {
   lua_getglobal(L, "tostring");
   lua_pushvalue(L, 2);
   lua_call(L, 1, 1);
-  return luaL_error(L, "attempt to assign property %s of mat4 (invalid property)", lua_tostring(L, -1));
+  return luaL_error(L, "attempt to assign property %s of Mat4 (invalid property)", lua_tostring(L, -1));
 }
 
 static int l_lovrMat4__index(lua_State* L) {
@@ -392,16 +435,20 @@ static int l_lovrMat4__index(lua_State* L) {
   lua_getglobal(L, "tostring");
   lua_pushvalue(L, 2);
   lua_call(L, 1, 1);
-  return luaL_error(L, "attempt to index field %s of mat4 (invalid property)", lua_tostring(L, -1));
+  return luaL_error(L, "attempt to index field %s of Mat4 (invalid property)", lua_tostring(L, -1));
 }
 
 const luaL_Reg lovrMat4[] = {
   { "equals", l_lovrMat4Equals },
   { "unpack", l_lovrMat4Unpack },
   { "getPosition", l_lovrMat4GetPosition },
+  { "setPosition", l_lovrMat4SetPosition },
   { "getOrientation", l_lovrMat4GetOrientation },
+  { "setOrientation", l_lovrMat4SetOrientation },
   { "getScale", l_lovrMat4GetScale },
+  { "setScale", l_lovrMat4SetScale },
   { "getPose", l_lovrMat4GetPose },
+  { "setPose", l_lovrMat4SetPose },
   { "set", l_lovrMat4Set },
   { "mul", l_lovrMat4Mul },
   { "identity", l_lovrMat4Identity },