+++ CHANGES.md
@@ -69,6 +69,7 @@ dev
 - Fix missing error in `CapsuleShape:setRadius/Length` and `CylinderShape:setRadius/Length`.
 - Fix issue where buffer memory wasn't recycled effectively.
 - Fix thread stack traces to use the correct filename.
+- Fix `t.headset.seated`.
 
 ### Deprecate